services/ai_service/requirements.txt
fastapi
uvicorn[standard]
openai
chromadb-client
httpx